Validation testing

+++ ValidationLibrary/Validator.cs
@@ -21,7 +21,6 @@ namespace ValidationLibrary
         /// The character sequence to terminate each line
         /// in the validation message.
         /// </summary>
-
         public static string LineEnd { get; set; } = "\n";
 
         #region Text Box
@@ -70,7 +69,7 @@ namespace ValidationLibrary
         /// <returns>True if textbox can be converted to an decimal, false and an error message if not.</returns>
         public static bool IsTextboxDecimal(string name, TextBox box)
         {
-            if (!int.TryParse(box.Text, out _))
+            if (!decimal.TryParse(box.Text, out _))
             {
                 MessageBox.Show(name + requiredDecimal, "Invalid Entry", MessageBoxButtons.OK, MessageBoxIcon.Warning);
                 box.Focus();
